Time: 25 ms (58.63%), Space: 86 MB (75.1%) - LeetHub

+++ 3634-minimum-removals-to-balance-array/3634-minimum-removals-to-balance-array.java
@@ -8,7 +8,7 @@ class Solution {
         int n=nums.length,i=0;
         int mini=n;
         for (int j = 0; j < nums.length; j++) {
-            while ((long) nums[j]>(long)nums[i]*k){
+            while (nums[j]>(long)nums[i]*k){
                 i++;
             }
            // System.out.println(n-j+i-1);
